fix aggiunto campi foto animali, eliminazioni migliorate, sincronizzazione calendar corretta

+++ app/routes/mungiture.js
@@ -5,6 +5,7 @@ import { assertAziendaOwnedByUser } from './aziende.js';
 import Azienda from '../models/azienda.js';
 import Animale from '../models/animale.js';
 import Mungitura from '../models/munigitura.js';
+import Lavorazione from '../models/lavorazione.js';
 import Sensore from '../models/sensore.js';
 import { ultimeLettureIot } from '../services/mqttService.js';
 
@@ -338,6 +339,17 @@ export const deleteMungitura = async (req, res) => {
         const ownershipCheck = await assertAziendaOwnedByUser(existingMungitura.aziendaId, req.user.userId);
         if (!ownershipCheck.ok) {
             return res.status(ownershipCheck.status || 403).json({ message: ownershipCheck.message });
+        }
+
+        const usedInLavorazione = await Lavorazione.exists({
+            aziendaId: existingMungitura.aziendaId,
+            'inputs.mungituraIds': existingMungitura._id
+        });
+
+        if (usedInLavorazione) {
+            return res.status(409).json({
+                message: 'Impossibile eliminare la mungitura: e gia utilizzata in una lavorazione'
+            });
         }
 
         await Mungitura.deleteOne({ _id: id });
